migrations/requirements.txt
GitPython==3.1.11
requests==2.25.0
